Refactor project structure: implement clean architecture, split UI into `main_app.py`, and move genetic algorithm logic to `solve_nqueens.py`. Update `readme.md` with new structure and principles. Append additional results to `Ganadores.txt`.
