+++ Ganadores.txt
@@ -626,3 +626,45 @@ Individuo 9836: [8, 5, 3, 6, 0, 7, 1, 4], Fitness: 0.0
 Individuo 9932: [4, 8, 5, 3, 1, 7, 2, 6, 0], Fitness: 0.0
 Individuo 9994: [5, 2, 0, 7, 3, 8, 6, 4, 1], Fitness: 0.0
 
+Individuo 48: [1, 7, 4, 6, 8, 0, 5, 3], Fitness: 0.0
+Individuo 61: [3, 7, 2, 8, 6, 4, 0, 5, 1], Fitness: 0.0
+Individuo 70: [2, 4, 8, 0, 5, 3, 1, 7, 6], Fitness: 0.0
+Individuo 152: [0, 6, 4, 7, 1, 8, 2, 5, 3], Fitness: 0.0
+Individuo 362: [6, 1, 7, 2, 0, 3, 8, 4, 5], Fitness: 0.0
+Individuo 446: [4, 0, 7, 3, 1, 8, 2, 5, 6], Fitness: 0.0
+Individuo 460: [2, 5, 8, 6, 3, 0, 7, 1], Fitness: 0.0
+Individuo 532: [3, 1, 4, 8, 0, 2, 7, 5], Fitness: 0.0
+Individuo 560: [6, 1, 7, 2, 0, 3, 8, 4, 5], Fitness: 0.0
+Individuo 572: [0, 6, 4, 7, 1, 8, 2, 5, 3], Fitness: 0.0
+Individuo 575: [4, 8, 0, 3, 6, 2, 5, 1, 7], Fitness: 0.0
+Individuo 643: [7, 2, 0, 5, 1, 8, 6, 3, 4], Fitness: 0.0
+Individuo 666: [2, 5, 1, 4, 7, 0, 6, 3], Fitness: 0.0
+Individuo 796: [8, 0, 5, 3, 1, 6, 4, 2], Fitness: 0.0
+Individuo 921: [3, 6, 0, 7, 1, 4, 2, 8], Fitness: 0.0
+
+Individuo 185: [7, 1, 4, 2, 0, 8, 3, 5], Fitness: 0.0
+Individuo 281: [3, 7, 2, 8, 6, 4, 0, 5, 1], Fitness: 0.0
+Individuo 351: [4, 6, 8, 2, 7, 1, 3, 5, 0], Fitness: 0.0
+Individuo 427: [3, 7, 2, 8, 6, 4, 0, 5], Fitness: 0.0
+Individuo 446: [6, 3, 0, 7, 1, 8, 5, 2, 4], Fitness: 0.0
+Individuo 487: [3, 7, 2, 8, 6, 4, 0, 5, 1], Fitness: 0.0
+Individuo 494: [3, 1, 6, 8, 2, 4, 7, 0], Fitness: 0.0
+Individuo 698: [3, 7, 2, 8, 6, 4, 0, 5, 1], Fitness: 0.0
+Individuo 765: [3, 7, 2, 8, 6, 4, 0, 5, 1], Fitness: 0.0
+Individuo 790: [0, 5, 8, 1, 3, 7, 2, 4, 6], Fitness: 0.0
+
+Individuo 315: [3, 1, 4, 8, 5, 0, 2, 6], Fitness: 0.0
+Individuo 508: [2, 7, 1, 3, 8, 6, 4, 0, 5], Fitness: 0.0
+Individuo 518: [1, 5, 8, 0, 7, 3, 6, 2, 4], Fitness: 0.0
+Individuo 535: [2, 7, 1, 3, 8, 6, 4, 0, 5], Fitness: 0.0
+Individuo 585: [4, 6, 3, 0, 2, 8, 5, 1], Fitness: 0.0
+Individuo 587: [3, 1, 4, 8, 5, 0, 2, 6], Fitness: 0.0
+Individuo 604: [6, 3, 0, 2, 8, 5, 1, 4, 7], Fitness: 0.0
+Individuo 662: [1, 5, 0, 6, 4, 2, 8, 3, 7], Fitness: 0.0
+Individuo 664: [6, 3, 0, 2, 8, 5, 1, 4, 7], Fitness: 0.0
+Individuo 769: [2, 8, 3, 0, 7, 1, 4, 6, 5], Fitness: 0.0
+Individuo 925: [3, 6, 0, 2, 8, 1, 7, 4], Fitness: 0.0
+Individuo 940: [0, 5, 1, 4, 7, 3, 8, 2], Fitness: 0.0
+Individuo 990: [4, 6, 3, 0, 7, 5, 8, 2], Fitness: 0.0
+Individuo 998: [3, 6, 8, 1, 4, 7, 0, 2], Fitness: 0.0
+
